chore : adding makefile - puts the app in build. Still some minor issues to solve prior to running on the target
