+++ led_control_project/main.c
@@ -1,0 +1,43 @@
+#include <stdio.h>
+
+#include "silly_led_control.h"
+
+int main(int argc, char** argv){
+    printf("hello there\n");
+    long duty   = 2;     // %
+    long period = 1000;  // ms
+    if (argc >= 2) period = atoi(argv[1]);
+    period *= 1000000;  // in ns
+
+    // compute duty period...
+    long p1 = period / 100 * duty;
+    long p2 = period - p1;
+
+    int led = open_led();
+    pwrite(led, "1", sizeof("1"), 0);
+
+    struct timespec t1;
+    clock_gettime(CLOCK_MONOTONIC, &t1);
+
+    int k = 0;
+    while (1) {
+        struct timespec t2;
+        clock_gettime(CLOCK_MONOTONIC, &t2);
+
+        long delta =
+            (t2.tv_sec - t1.tv_sec) * 1000000000 + (t2.tv_nsec - t1.tv_nsec);
+
+        int toggle = ((k == 0) && (delta >= p1)) | ((k == 1) && (delta >= p2));
+        if (toggle) {
+            t1 = t2;
+            k  = (k + 1) % 2;
+            if (k == 0)
+                pwrite(led, "1", sizeof("1"), 0);
+            else
+                pwrite(led, "0", sizeof("0"), 0);
+        }
+    }
+
+    printf("have a nice day\n");
+    return 0;
+}
