+++ led_control_project/src/silly_led_control.c
@@ -1,0 +1,48 @@
+/**
+ * Copyright 2018 University of Applied Sciences Western Switzerland / Fribourg
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Project: HEIA-FR / HES-SO MSE - MA-CSEL1 Laboratory
+ *
+ * Abstract: System programming -  file system
+ *
+ * Purpose: NanoPi silly status led control system
+ *
+ * Author:  Daniel Gachet
+ * Date:    07.11.2018
+ */
+
+#include "silly_led_control.h"
+
+int open_led(){
+    // unexport pin out of sysfs (reinitialization)
+    int f = open(GPIO_UNEXPORT, O_WRONLY);
+    write(f, LED, strlen(LED));
+    close(f);
+
+    // export pin to sysfs
+    f = open(GPIO_EXPORT, O_WRONLY);
+    write(f, LED, strlen(LED));
+    close(f);
+
+    // config pin
+    f = open(GPIO_LED "/direction", O_WRONLY);
+    write(f, "out", 3);
+    close(f);
+
+    // open gpio value attribute
+    f = open(GPIO_LED "/value", O_RDWR);
+    return f;
+}
+
